Add game-state and websocket communication
A websocket is connected when the game page loads. On initialisation
player is given a random location, which is sent back to the server via
websocket, and an object is added to game state to represent player.

Need to add some server->client communication to send game state updates
so that the player can see other objects' position/movment.